+++ src/main/java/com/solvd/construction/model/Country.java
@@ -1,17 +1,17 @@
 package com.solvd.construction.model;
 
 public class Country {
-    private final int id;
+    private final long id;
     private String countryName;
     private String postalCode;
 
-    public Country(int id, String countryName, String postalCode) {
+    public Country(long id, String countryName, String postalCode) {
         this.id = id;
         this.countryName = countryName;
         this.postalCode = postalCode;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
